backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
chromadb==0.4.22
httpx==0.26.0
openai==1.10.0
ollama==0.1.2
Pillow==10.2.0
python-docx==1.1.0
pypdf==4.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
watchfiles==0.21.0
